refactor(chat): extract event-preview helpers into a pure util
Move getEventReactionPreview, formatReplyTooltipText, and the
EVENT_REACTION_PREVIEW_COLORS palette out of MessageDisplay.jsx into
src/utils/messageDisplayHelpers.js. These are MessageDisplay-only and
non-JSX (getEventReactionPreview returns plain objects whose Icon/
trailingIcon fields are lucide component references), so the util imports
parseSystemMessage and the needed lucide icons. No behavior change —
verbatim move plus the two exports. Smoke-verified: event reaction
previews and reply tooltips render correctly.

MessageDisplay drops from 4597 to 4367 lines.

Stage 2 of the MessageDisplay.jsx decomposition (Tier 2).
